+++ config/config.go
@@ -3,6 +3,8 @@ package config
 import (
 	"flag"
 	"github.com/NubeIO/configor"
+	"github.com/NubeIO/flow-framework/utils/file"
+	"github.com/NubeIO/flow-framework/utils/security"
 	"path"
 )
 
@@ -38,12 +40,12 @@ type Configuration struct {
 			PluginsDir        string `default:"plugins"`
 			UploadedImagesDir string `default:"images"`
 		}
-		DeviceInfoFile       string `default:"/data/rubix-registry/device_info.json"`
-		AuthDataDir          string `default:"/data/rubix-service"`
-		RelativeAuthDataFile string `default:"/data/internal_token.txt"`
+		DeviceInfoFile    string `default:"/data/rubix-registry/device_info.json"`
+		TokenFolder       string `default:"/data/rubix-service/data/"`
+		InternalTokenFile string `default:"internal_token.txt"`
 	}
-	Prod            bool `default:"false"`
-	Auth            bool `default:"false"`
+	Prod            bool  `default:"false"`
+	Auth            *bool `default:"true"`
 	ProducerHistory struct {
 		Cleaner struct {
 			Enable              *bool `default:"true"`
@@ -55,6 +57,7 @@ type Configuration struct {
 			SyncPeriod int   `default:"10"`
 		}
 	}
+	SecretKey string
 }
 
 var config *Configuration = nil
@@ -66,6 +69,7 @@ func Get() *Configuration {
 func CreateApp() *Configuration {
 	config = new(Configuration)
 	config = config.Parse()
+	config = config.HandleSecretKey()
 	err := configor.New(&configor.Config{EnvironmentPrefix: "FLOW"}).Load(config, path.Join(config.GetAbsConfigDir(), "config.yml"))
 	if err != nil {
 		panic(err)
@@ -88,6 +92,16 @@ func (conf *Configuration) Parse() *Configuration {
 	return conf
 }
 
+func (conf *Configuration) HandleSecretKey() *Configuration {
+	secretKey, _ := file.ReadFile(path.Join(config.GetAbsConfigDir(), "secret.txt"))
+	if secretKey == "" {
+		secretKey = security.GenerateToken()
+		_, _ = file.WriteDataToFileAsString(path.Join(config.GetAbsConfigDir(), "secret.txt"), secretKey)
+	}
+	conf.SecretKey = secretKey
+	return conf
+}
+
 func (conf *Configuration) GetAbsDataDir() string {
 	return path.Join(conf.Location.GlobalDir, conf.Location.DataDir)
 }
@@ -102,4 +116,8 @@ func (conf *Configuration) GetAbsPluginDir() string {
 
 func (conf *Configuration) GetAbsUploadedImagesDir() string {
 	return path.Join(conf.GetAbsDataDir(), conf.Location.Data.UploadedImagesDir)
+}
+
+func (conf *Configuration) GetAbsInternalTokenFile() string {
+	return path.Join(conf.Location.TokenFolder, conf.Location.InternalTokenFile)
 }