+++ database/sync_flow_network.go
@@ -11,7 +11,7 @@ import (
 
 func (d *GormDatabase) SyncFlowNetwork(body *model.FlowNetwork) (*model.FlowNetworkClone, error) {
 	if !boolean.IsTrue(body.IsMasterSlave) {
-		if boolean.IsTrue(body.IsRemote) {
+		if boolean.IsTrue(body.IsRemote) && boolean.IsFalse(body.IsTokenAuth) {
 			accessToken, err := client.GetFlowToken(*body.FlowIP, *body.FlowPort, *body.FlowUsername, *body.FlowPassword)
 			if err != nil {
 				return nil, err
@@ -20,6 +20,10 @@ func (d *GormDatabase) SyncFlowNetwork(body *model.FlowNetwork) (*model.FlowNetw
 		}
 	}
 	cli := client.NewFlowClientCliFromFN(body)
+	// In this case it's even though FN, this is clone side & need to communicate with slave device from master
+	if boolean.IsTrue(body.IsMasterSlave) {
+		cli = client.NewMasterToSlaveSession(body.GlobalUUID)
+	}
 	remoteDeviceInfo, err := cli.DeviceInfo()
 	if err != nil {
 		return nil, err